Move all of button into remap controls
