+++ src/Compatibility/Core/src/AppHostBuilderExtensions.cs
@@ -179,6 +179,7 @@ namespace Microsoft.Maui.Controls.Hosting
 					// Update the mappings for IView/View to work specifically for Controls
 					VisualElement.RemapForControls();
 					Label.RemapForControls();
+					Button.RemapForControls();
 
 				})
 				.ConfigureServices<MauiCompatBuilder>();
