+++ src/Controls/src/Core/Platform/Standard/Extensions/ButtonExtensions.cs
@@ -1,0 +1,13 @@
+using System;
+using Microsoft.Maui.Handlers;
+using static Microsoft.Maui.Controls.Button;
+
+namespace Microsoft.Maui.Controls.Platform
+{
+	public static class ButtonExtensions
+	{
+		public static void UpdateContentLayout(this object nativeButton, Button button)
+		{
+		}
+	}
+}
