--- src/Core/src/Core/IButtonContentLayout.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Microsoft.Maui
-{
-	internal interface IButtonContentLayout
-	{
-		ButtonContentLayout ContentLayout { get; }
-	}
-}
