+++ src/Core/src/Handlers/Button/ButtonHandler.Standard.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace Microsoft.Maui.Handlers
 {
-	public partial class ButtonHandler : ViewHandler<IButton, object>
+	public sealed partial class ButtonHandler : ViewHandler<IButton, object>
 	{
 		protected override object CreateNativeView() => throw new NotImplementedException();
