+++ src/Core/src/Platform/Android/ButtonExtensions.cs
@@ -34,56 +34,5 @@ namespace Microsoft.Maui
 				(int)padding.Right,
 				(int)padding.Bottom);
 		}
-
-		public static void UpdateContentLayout(this MaterialButton materialButton, IButton button)
-		{
-			var context = materialButton.Context;
-			if (context == null)
-				return;
-
-			var icon = materialButton.Icon ??
-						TextViewCompat.GetCompoundDrawablesRelative(materialButton)[3];
-
-			if (icon != null &&
-				!String.IsNullOrEmpty(button.Text) &&
-				button is IButtonContentLayout cl)
-			{
-				var contentLayout = cl.ContentLayout;
-
-				// IconPadding calls materialButton.CompoundDrawablePadding				
-				// Which is why we don't have to worry about calling setCompoundDrawablePadding
-				// ourselves for our custom implemented IconGravityBottom
-				materialButton.IconPadding = (int)context.ToPixels(contentLayout.Spacing);
-
-				switch (contentLayout.Position)
-				{
-					case ButtonContentLayout.ImagePosition.Top:
-						materialButton.Icon = icon;
-						materialButton.IconGravity = MaterialButton.IconGravityTop;
-						break;
-					case ButtonContentLayout.ImagePosition.Bottom:
-						materialButton.Icon = null;
-						TextViewCompat.SetCompoundDrawablesRelative(materialButton, null, null, null, icon);
-						materialButton.IconGravity = MauiMaterialButton.IconGravityBottom;
-						break;
-					case ButtonContentLayout.ImagePosition.Left:
-						materialButton.Icon = icon;
-						materialButton.IconGravity = MaterialButton.IconGravityStart;
-						break;
-					case ButtonContentLayout.ImagePosition.Right:
-						materialButton.Icon = icon;
-						materialButton.IconGravity = MaterialButton.IconGravityEnd;
-						break;
-				}
-			}
-			else
-			{
-				// Don't remove this otherwise the button occasionally measures wrong
-				// on first load
-				materialButton.Icon = icon;
-				materialButton.IconPadding = 0;
-				materialButton.IconGravity = MaterialButton.IconGravityTextStart;
-			}
-		}
 	}
 }
